made mvnormpdf more robust to high-dimensional distributions with significant spread of data in only a few dimensions
The problem which occurred: I had a covariance matrix with 90 dimensions, but onlyabout 6 dimensions contained most of the variance of the data. Matlab computed the determinant of that matrix to be 0, but it was a proper covariance matrix with all eigenvalues > 0, even though more than 80 of them were 1e-4 or so.

The fix: When the actual density value is supposed to be computed, nothing can be done. The function will just return NaN. But for the log-density you can use the log-determinant to stay within the numerical limits. Hence, I now compute with log-determinant. I still use Matlab's det function to get the determinant first which I assume to be more efficient than my own attempt, but when det returns 0, I compute the eigenvalues of the covariance explicitly and compute the log-determinant from them by summing their log-values.