fixed reordering issues for std shades
In the old version a newly plotted std shade was moved behind an existing one.
I have now changed this, moving all previously existing std shades into the
background in their predefined order. This behaviour is consistent with the
rest of Matlab's plotting functions (new objects are plotted on top).