accounted for determinant = Inf for large matrices
This happens when the determinant cannot be represented numerically, because it
is so large. I don't need to represent the determinant, however, because I'm
only interested in the log-determinant. Hence, when the determinant turns out
to be Inf, I compute it via the sum of log-eigenvalues. I keep first computing
the determinant itself, because I assume that it is more efficient to compute
the determinant than the eigenvalues. This needs to be checked perhaps.